add to_function

+++ src/deepquantum/mbqc/command.py
@@ -154,7 +154,7 @@ class Measurement(Command):
             # M^{YZ,α} X^s Z^t = M^{YZ,(-1)^t·α+(s+t)π)}
         cir = QubitCircuit(nqubit=nqubit)
         cir.j(wires=wire, plane=self.plane, encode=True)
-        final_state = cir(data=alpha, state=init_state.squeeze(0))
+        final_state = cir(data=alpha.to(self.angle.dtype), state=init_state.squeeze(0))
         rst = cir.measure(shots=1, wires=wire)
         state = []
         if isinstance(rst, list):
